update change contracts

+++ contracts/complex/events.sol
@@ -4,9 +4,9 @@ pragma solidity >=0.8.2 <0.9.0;
 
 
 contract Event {
-   event TransferTokens(bytes32 fromWallet, bytes32 toWallet, uint256 amount);
+   event TransferTokens(bytes32 fromWallet, bytes32 toWallet, uint128 amount);
 
-   function emitTransfer(bytes32 fromWalletId, bytes32 toWalletId, uint256 amount) public {
+   function emitTransfer(bytes32 fromWalletId, bytes32 toWalletId, uint128 amount) public {
         emit TransferTokens(fromWalletId, toWalletId, amount);
     }
 }
